@@ -10,14 +10,14 @@
     "buffer": "^3.5.4",
     "jsrsasign": "5.0.1",
     "keymirror": "~0.1.0",
-    "react": "^0.14.7",
-    "react-native": "0.22.2",
+    "react": "^15.1.0",
+    "react-native": "0.27.2",
     "react-native-button": "*",
-    "react-native-vector-icons": "^1.3.3",
+    "react-native-vector-icons": "^2.0.3",
     "react-redux": "^4.4.5",
     "redux": "^3.4.0",
     "redux-logger": "^2.6.1",
-    "redux-saga": "^0.10.0",
+    "redux-saga": "^0.10.5",
     "redux-storage": "^4.0.0",
     "redux-storage-decorator-filter": "^1.1.4",
     "redux-storage-engine-reactnativeasyncstorage": "^1.0.0",
